pyproject.toml
[tool.poetry]
name = "repl_python3_Basic-Bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
replit = "^3.1.0"
flask = "^2.0.1"
python-dotenv = "^0.17.1"
click = "^8.0.1"
SQLAlchemy = "^1.4.17"
aiosqlite = "^0.17.0"
Jinja2 = "^3.0.1"
pyleri = "^1.3.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
